46.2. useTransition in a Neat, Reusable Module
The skipHiddenProperty Value
Anyways, back at the browser, type to re-open the suggestions then click off of to close it. That happened instantly! Where was our transition?
Inspect the element. Ah: see that little hidden attribute on the results div? That was added by the stimulus-autocomplete controller as soon as we clicked off. Thanks that, the element became hidden instantly instead of waiting for our transition.
Normally, that's great! It's how stimulus-autocomplete hides the results. But now that we are controlling the hiding and showing with our transition behavior, we do not want this hidden attribute to be added. Fortunately, assuming my PR is merged, we can pass a value to disable that behavior.
In the template, on the autocomplete controller, pass a new value called skipHiddenProperty set to true.
That literally says: please do not set that hidden property: we are handling the hiding and showing ourselves.
Let's try it out again. I'll type... we still get the nice fade in... and when I click off. Yes! It fades out!
And.... we're done! I mean, the whole tutorial is done! I hope you found this journey through Stimulus as refreshing as I did. I love coding with Stimulus.
In the next tutorial in the series - about Turbo - I hope to show prove that we can have an even more dynamic and speedy app while writing even less custom JavaScript.
Let us know what cool stuff your building. And, as always, if you have any questions, we're here for you in the comments section.
Okay, friends, seeya next time!